WIP: Multi-language support proposal
